Turtle Temperature-Dependent Sex Determination Data

This dataset, Tsd_turtle, is a data frame containing extensive information on temperature-dependent sex determination (TSD) in reptiles. The dataset includes 2,738 observations and 48 variables, covering species, location, incubation temperatures, egg and hatchling measurements, sex ratios, clutch information, and references. All values are preserved as in the original source, with no modifications.

Usage

data(Tsd_turtle)

Format

A data frame with 2,738 observations and 48 variables:

Species

Species name (factor)

Country

Country of observation (factor)

Area

Specific area or locality (character string)

Longitude

Longitude of the location (numeric)

Latitude

Latitude of the location (numeric)

Subspecies

Subspecies information (character string)

RMU.2010

RMU code (2010 version) (factor)

RMU.2023

RMU code (2023 version) (factor)

Incubation.temperature.set

Incubation temperature set experimentally (numeric)

Incubation.temperature.recorded

Recorded incubation temperature (numeric)

Incubation.temperature.corrected

Corrected incubation temperature (numeric)

Duplicated.data

Indicator for duplicated data (logical)

Duplicate

Duplicate identifier (character string)

Incubation.temperature.Constant

Constant temperature indicator (logical)

Incubation.temperature.Accuracy

Accuracy of incubation temperature (numeric)

Incubation.temperature.SD

Standard deviation of incubation temperature (numeric)

Incubation.temperature.Amplitude

Amplitude of incubation temperature (numeric)

2ndThird.Incubation.temperature.Amplitude

Amplitude during 2nd/3rd incubation period (numeric)

Correction.factor

Correction factor applied (numeric)

Egg.mass.mean

Mean egg mass (numeric)

Egg.mass.sd

Standard deviation of egg mass (numeric)

IP.min

Minimum incubation period (numeric)

IP.max

Maximum incubation period (numeric)

IP.mean

Mean incubation period (numeric)

IP.SD

Standard deviation of incubation period (numeric)

IP.SE

Standard error of incubation period (numeric)

Length.hatchlings.mean

Mean length of hatchlings (numeric)

Length.hatchlings.sd

Standard deviation of hatchling length (numeric)

SCL.hatchlings.mean

Mean straight carapace length of hatchlings (numeric)

SCL.hatchlings.sd

Standard deviation of hatchling carapace length (numeric)

Mass.hatchlings.mean

Mean hatchling mass (numeric)

Mass.hatchlings.sd

Standard deviation of hatchling mass (numeric)

Total

Total eggs recorded (numeric)

Hatched

Number of hatched eggs (numeric)

NotHatched

Number of unhatched eggs (numeric)

Undeveloped

Number of undeveloped eggs (numeric)

Unidentified

Number of unidentified eggs (numeric)

Intersexes

Number of intersex hatchlings (numeric)

Males

Number of male hatchlings (numeric)

Females

Number of female hatchlings (numeric)

Sexed

Number of sexed hatchlings (numeric)

Number.clutch

Number of clutches observed (numeric)

Clutch

Clutch identifier (factor)

Box

Box or container identifier (character string)

Reference

Reference for the data (character string)

Note

Additional notes (character string)

Digital_Identifier

Digital identifier (character string)

Version

Dataset version (Date)

Source

Data taken from the embryogrowth package version 10.3

Details

The dataset name has been kept as 'Tsd_turtle' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the BioWorldR package and assists users in identifying its specific characteristics. The suffix 'df' is not used here because the dataset is identified by its function within BioWorldR. The original content has not been modified.